add post Disease History

+++ app/src/main/java/com/example/rifsa_mobile/model/entity/remote/disease/DiseaseResultResponse.kt
@@ -24,9 +24,9 @@ data class DiseaseResultDataResponse(
  @SerializedName("indikasi")
  val indikasi: String,
  @SerializedName("latitude")
- val latitude: Int,
+ val latitude: Double,
  @SerializedName("longitude")
- val longitude: Int,
+ val longitude: Double,
  @SerializedName("nama")
  val nama: String,
  @SerializedName("tanggal")
